added folder checks

+++ editor/pages/delete.php
@@ -27,17 +27,40 @@ if (isset($_POST['delete_button'])) {
             $path_for_website_unlink = realpath("../../")."/website/media/images/".end($exploded_path);
             $path_for_editor_unlink = realpath("../../") . "/editor/media/images/".end($exploded_path);
             
+            $deleted_from_website = false;
+            $deleted_from_editor = false;
+
             // Delete files
-            unlink($path_for_website_unlink);
-            unlink($path_for_editor_unlink);
+            if (unlink($path_for_website_unlink)) {
+                echo "success";
+                $deleted_from_website = true; // flag for web side
+            }
+            else {
+                echo "Failed deleting from website folder";
+            }
 
-            // Delete from DB
-            $sql = "DELETE FROM photo WHERE photo_id = $image";
-            $stmt = $GLOBALS['pdo']->prepare($sql);
-            $stmt->execute();
+            if (unlink($path_for_editor_unlink)) {
+                echo "success";
+                $deleted_from_editor = true; // flag for editor side
+            }
+            else {
+                echo "Failed deleting from editor folder";
+            }
 
-            // Refresh Page with success messages
-            header("Location: edit_carousel_photos.php?deletesuccess");
+            if ($deleted_from_editor && $deleted_from_website) {
+
+                // Delete from DB
+                $sql = "DELETE FROM photo WHERE photo_id = $image";
+                $stmt = $GLOBALS['pdo']->prepare($sql);
+                $stmt->execute();
+
+                // Refresh Page with success messages
+                header("Location: edit_carousel_photos.php?deletesuccess");
+                } 
+            
+                else {
+                    echo "Failed to delete from database";
+                }
         }
     } else {
         // Refresh Page with error messages
